Minor Changes
I think to account for increased and decreasing f in min heap, we just put a sink function after the swim (if its in the right place it wont sink, right, so it wont undo what swim did) i changed this in heap
We need to implement the binary heap with the <g as well
returning 0 or 1 fixes the NoneType doesn’t have parent attribute 
There was an error in adaptive, i fixed that
For the visuals, they are make the program go slower, if you want to speed it up, don’t update the window as often, tho its prettier. updates are done using w.master.update(), you can move it outside the while loops, so the changes only display then, depends on what you want to see. 
If you figure out the u-turns, please add that to AdaptiveMain also 
Adding visualization to backward, will be the same as the changes i made to forward/adaptive